+++ src/main/java/net/jwarren/workers/misc/WorkerNotFoundException.java
@@ -1,7 +1,7 @@
 package net.jwarren.workers.misc;
 
 public class WorkerNotFoundException extends RuntimeException {
-    public WorkerNotFoundException(Long id) {
+    public WorkerNotFoundException(Integer id) {
         super("Could not retrieve worker with ID " + id);
     }
 }
